Bugfixing for pubsystem's !buy related autobots. - Baseterr time lowered to two minutes, as stated in the help. - Baseterr and roofturret communication between pubsystem and pubautobot now fully happens over IPC. This to prevent exploits. - Added more safety measures in the autobot's code, but more unexpected errors remain. - Changed flagsaver to actually claim the flag for the purchaser. It is now claimed instantly.
NOTES:
- Need to confirm with fiS/WingZero if flagsaver is intended to claim the flag instantly or obey the 3-second grace period.
- Need to confirm with fiS/WingZero if baseterr time needs to be two or five minutes. (Or something completely different.)
WingZero:
Despite fixing flagsaver with the context setup, the bot still needs to send a packet to the server to make it function correctly. This has to do with the server not correctly registering who owns the flag, and thus who to assign the 200 points and orange/blue color to. Bot/pubsystem-wise the context method did suffice though.
Future pubautobot developers:
Please note that a position packet from a bot on a flag does not trigger a flag capture packet and thus does not trigger a flag claimed event. (This might also be true for a few other event types.)

git-svn-id: http://twcore.org/svn/trunk@7753 9f8b3893-2a26-0410-9d00-a71e40fd2f77