Removing debug code for a long-standing ArrayOutOfBoundsException that ThePAP was never able to figure out, and neither was I.
Issue: indexOf seems to return an index that is less than -1, even when it should never do this (according to its docs). This bad index is then passed to substring, causing the issue. I'm guessing this is actually a bug with indexOf(char) in String.java. Have reviewed String.java code but didn't see anything pop out. May have to do with special unhandled characters, which is what ThePAP's modifications to Message.java were for to begin with: allowing some non-standard unicode to get through without causing problems for TWCore.

Anyhow, the easy fix for this is just checking the index ourselves manually and not parsing the message if we can't read a name and msg from it. No big deal.

This should reduce CPU load on the bots a little bit, because string concats are expensive, and this was forcing every bot to do 2-3 unnecessary concats for every single message received.

git-svn-id: http://twcore.org/svn/trunk@8994 9f8b3893-2a26-0410-9d00-a71e40fd2f77